btrfs: test btrfs specific fsverity corruption
There are some btrfs specific fsverity scenarios that don't map
neatly onto the tests in generic/574 like holes, inline extents,
and preallocated extents. Cover those in a btrfs specific test.

This test relies on the btrfs implementation of fsverity in:
<kernel patches>
and it relies on btrfs-corrupt-block for corruption, with the patches:
<corrupt block patches>

Signed-off-by: Boris Burkov <boris@bur.io>